debugging manager side

+++ bamazon.sql
@@ -9,15 +9,15 @@ CREATE TABLE product (
   item_id INT NOT NULL AUTO_INCREMENT,
   product_name VARCHAR(100) NULL,
     department_name VARCHAR(100) NULL,
-  price INTEGER(10) NULL,
+  price DECIMAL(6,2) NULL,
     stock_quantity INTEGER(10) NULL,
-    product_sales INTEGER(10) NULL,
+    product_sales DECIMAL(6,2) NULL,
   PRIMARY KEY (item_id)
 );
 
 CREATE TABLE departments (
 department_id INT NOT NULL AUTO_INCREMENT,
 department_name VARCHAR(100) NULL,
-overhead_costs INTEGER(10) NULL,
+overhead_costs DECIMAL(6,2) NULL,
 PRIMARY KEY (department_id)
 );
